AG-13304 - Fix circular dependency.
